+++ push_swap/push_swap.h
@@ -6,7 +6,7 @@
 /*   By: trpham <trpham@student.hive.fi>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/13 10:41:55 by trpham            #+#    #+#             */
-/*   Updated: 2025/01/29 20:02:19 by trpham           ###   ########.fr       */
+/*   Updated: 2025/01/30 17:18:27 by trpham           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,24 +24,12 @@ typedef struct s_node
 	int				content;
 	struct s_node	*next;
 	struct s_node	*prev;
-
 	int				index;
-	int				moves_to_top; // moves required to bring to top
+	int				moves_to_top;
 	struct s_node *best_friend;
 	int	friend_moves_to_top;
 	int	total_moves;
 }	t_node;
-
-// typedef struct s_stack
-// {
-// 	t_node	*head;
-// 	t_node	*end;
-// 	t_node	**arr;
-// 	int	current_level;
-// 	int	stack_count;
-// }	t_stack;
-
-
 
 void	swap_stack(t_node **lst, char	c);
 void 	rotate_stack(t_node **lst, char c);
@@ -50,9 +38,6 @@ void	push_stack(t_node **stack_1, t_node **stack_2, char	c);
 void	reverse_both(t_node **stack_1, t_node	**stack_2);
 void	rotate_both(t_node **stack_1, t_node	**stack_2);
 void	swap_both(t_node **stack_1, t_node	**stack_2);
-
-
-
 
 int			node_lst_size(t_node *lst);
 void		print_list(t_node *lst);
@@ -74,6 +59,17 @@ void	sort_stack_of_five(t_node **stack_a, t_node **stack_b, int stack_size);
 void	min_value_to_top_sort(t_node **stack_a, t_node **stack_b);
 void	ft_sort(t_node **stack_a, t_node **stack_b);
 int		min_node(t_node **stack_a);
+
+int	*stack_dup(t_node	*stack_a, int size);
+int	find_mean(t_node *stack, int size);
+void	partition_by_mean(t_node **stack_a, t_node **stack_b, int size);
+int	calculate_moves_to_top(t_node *stack, t_node *node);
+int	find_index(t_node *stack, t_node *node);
+void	execute_best_move(t_node **stack_a, t_node **stack_b);
+void	move_to_top(t_node **stack, t_node *node, char c);
+t_node	*find_best_friend(t_node *stack_a, int nb);
+void update_move_info(t_node *stack_a, t_node *stack_b);
+t_node	*find_min_node(t_node *stack);
 // void	quick_sort(t_node	**stack_a, t_node	**stack_b);
 
 #endif
